+++ CHANGELOG.md
@@ -6,7 +6,8 @@
       - the search is an OR operation meaning that at least one tag must match.
       - the search is case-sensitive.
     - `result_key` can now be set to `*` which returns the entire record for each match including metadata like `tags` and `category`.
-- Added the `register_task_templates()` method to `functions`.
+- Moved all `register_*` methods out of `functions.py` and into `registry.py`. 
+- Added the `register_task_templates()`
   - This method registers the contents of YAML files stored in CloudHarvest `template` directories throughout the application and its packages.
   - The `cls` is populated by the actual content of the file, a dictionary.
 - `Registry.add()` now registers objects using a `category-task` convention